+++ R/col-transition_risk_profile.R
@@ -1,0 +1,35 @@
+col_uuid <- function() {
+  "activity_uuid_product_uuid"
+}
+
+col_co2_footprint <- function() {
+  "co2_footprint"
+}
+
+col_type <- function() {
+  "type"
+}
+
+col_sector <- function() {
+  "sector"
+}
+
+col_subsector <- function() {
+  "subsector"
+}
+
+col_year <- function() {
+  "year"
+}
+
+col_scenario <- function() {
+  "scenario"
+}
+
+col_targets <- function() {
+  "reductions"
+}
+
+col_ranking <- function() {
+  "profile_ranking"
+}