+++ R/tiltTransitionRisk-package.R
@@ -2,5 +2,24 @@
 "_PACKAGE"
 
 ## usethis namespace: start
+#' @importFrom dplyr .data
+#' @importFrom dplyr all_of
+#' @importFrom dplyr distinct
+#' @importFrom dplyr filter
+#' @importFrom dplyr full_join
+#' @importFrom dplyr mutate
+#' @importFrom dplyr select
+#' @importFrom glue glue
+#' @importFrom purrr walk
+#' @importFrom readr read_csv
+#' @importFrom rlang abort
+#' @importFrom stats quantile
+#' @importFrom tibble tribble
+#' @importFrom tiltIndicator epa_compute_profile_ranking
+#' @importFrom tiltIndicator spa_compute_profile_ranking
+#' @importFrom tiltToyData toy_emissions_profile_products_ecoinvent
+#' @importFrom tiltToyData toy_sector_profile_any_scenarios
+#' @importFrom tiltToyData toy_sector_profile_companies
+#' @importFrom utils write.csv
 ## usethis namespace: end
 NULL