+++ NEWS.md
@@ -1,5 +1,12 @@
 <!-- NEWS.md is maintained by https://cynkra.github.io/fledge, do not edit -->
 
+
+
+# tiltTransitionRisk 0.0.0.9002 (2024-06-03)
+
+* New `add_thresholds_transition_risk()` function adds low and high thresholds for 
+transition risk scores (#13).
+
 # tiltTransitionRisk 0.0.0.9001 (2024-06-03)
 
 * Add NEWS.md.